--- src/control.js
@@ -1,6 +1,0 @@
-export const accesoGmail = (email, password) => {
-  return localStorage.setItem().signInWithEmailAndPassword(email, password);
-}
-export const registroAcceso = (email, password) => {
-  return localStorage.setItem().createUserWithEmailAndPassword(nombre, correo, contraseña);
-}
