--- src/main.js
@@ -1,4 +1,0 @@
-// Este es el punto de entrada de tu aplicacion
-import { myFunction } from './lib/index.js';
-
-myFunction();
