--- src/validar.js
@@ -1,5 +1,0 @@
-// llenarRegistro();
-
-// const llenandoRegistro = () =>{
-
-// }
